Replace JUnit artifact upload with test results summary
Removed the action for saving JUnit results as an artifact in the event of failure. Instead, a new action has been added to output test results as a summary irrespective of success or failure. This should provide an easy and clear understanding of test results every time an action runs.